+++ lib/pow/phoenix/mailer/template.ex
@@ -11,6 +11,8 @@ defmodule Pow.Phoenix.Mailer.Template do
         template :mail, "Subject line", "Text content", "<p>HTML content</p>"
       end
   """
+
+  @doc false
   defmacro __using__(_opts) do
     quote do
       import unquote(__MODULE__)