+++ lib/pow/phoenix/mailer/view.ex
@@ -2,7 +2,16 @@ defmodule Pow.Phoenix.Mailer.View do
   @moduledoc """
   View macros for `Pow.Phoenix.Mailer` that calls render methods generated with
   `Pow.Phoenix.Mailer.Template`.
+
+  ## Usage
+
+      defmodule MyExtension.Phoenix.MailerView do
+        @moduledoc false
+        use Pow.Phoenix.Mailer.View
+      end
   """
+
+  @doc false
   defmacro __using__(_opts) do
     quote do
       @template_module Pow.Phoenix.View.__template_module__(__MODULE__)