+++ lib/pow/phoenix/view.ex
@@ -2,6 +2,13 @@ defmodule Pow.Phoenix.View do
   @moduledoc """
   View macros for Pow Phoenix, that calls render methods generated with
   `Pow.Phoenix.Template`.
+
+  ## Usage
+
+      defmodule MyExtension.Phoenix.CustomView do
+        @moduledoc false
+        use Pow.Phoenix.View
+      end
   """
 
   @doc false
@@ -16,21 +23,20 @@ defmodule Pow.Phoenix.View do
   end
 
   @doc false
-  @spec __template_module__(atom()) :: atom()
   def __template_module__(view_module) do
-    [name | rest] =
+    [view_module | context] =
       view_module
       |> Module.split()
       |> Enum.reverse()
 
-    name =
-      name
+    template_module =
+      view_module
       |> String.trim_trailing("View")
       |> Kernel.<>("Template")
 
-    rest
+    context
     |> Enum.reverse()
-    |> Enum.concat([name])
+    |> Enum.concat([template_module])
     |> Module.concat()
   end
 end