+++ test/pow/extension/phoenix/messages_test.exs
@@ -8,15 +8,15 @@ defmodule Pow.Extension.Phoenix.MessagesTest do
     use Pow.Extension.Phoenix.Messages,
       extensions: [Pow.Extension.Phoenix.MessagesTest]
 
-    def pow_a(_conn), do: "Overridden"
+    def pow_extension_phoenix_messages_test_a(_conn), do: "Overridden"
   end
 
   use ExUnit.Case
   doctest Pow.Extension.Phoenix.Messages
 
   test "can override messages" do
-    assert Messages.pow_a(nil) == "Overridden"
-    assert Messages.pow_b(nil) == "Second"
+    assert Messages.pow_extension_phoenix_messages_test_a(nil) == "Overridden"
+    assert Messages.pow_extension_phoenix_messages_test_b(nil) == "Second"
   end
 
   test "has fallback module" do